# docs/contracts/openapi.foundation.yaml
openapi: 3.1.0
info:
  title: Foundation Core API
  version: 1.0.0
  description: Reusable backend-first runtime API for screenshot-driven frontend templates.
servers:
  - url: https://foundation-core.example.com
paths:
...
  /api/forms/{formId}/submit:
    post:
      summary: Submit a validated lead form
      parameters:
        - in: path
          name: formId
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Lead accepted
        '400':
          description: Invalid payload
        '429':
          description: Rate limit exceeded
        '422':
          description: Spam or rejected form
...